requirements.txt
fastapi
fastapi-middleware-cors
transformers
Pillow
torch
numpy
uvicorn
